Removed call to adjustTime() between calls to setTime() and hour().
This fixes occasional failure to reflect time zone (HoursOffset)
after updating the time from GPS.  The Time library contains a
refreshCache() function that updates the cache only if the time
has changed since the last cache update.

adjustTime does not call refreshCache().  hour() does call it, with
the current time.  If the current time is the same as the cached time
from the setTime() call, then the cache of tm is not updated to reflect
the adjustTime() call.  Then, the timezone adjustment is ignored.